+++ core/constants.php
@@ -15,8 +15,8 @@ if (PHP_SAPI !== 'cli' && !headers_sent()) {
     header('Referrer-Policy: strict-origin-when-cross-origin');
 }
 
-define('SNAPSMACK_VERSION', 'Alpha 0.7.33');
-define('SNAPSMACK_VERSION_SHORT', '0.7.33');
+define('SNAPSMACK_VERSION', 'Alpha 0.7.34');
+define('SNAPSMACK_VERSION_SHORT', '0.7.34');
 define('SNAPSMACK_VERSION_CODENAME', 'Perch');
 
 // --- VERSION COMPARISON ---